Temporarily moved blocks pkg back to repo root for diffing
