+++ blocks/common.go
