+++ blocks/common_test.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/facebookincubator/TTP-Runner/pkg/blocks"
+	"github.com/facebookincubator/TTP-Runner/blocks"
 	goutils "github.com/l50/goutils"
 	"github.com/stretchr/testify/assert"
 )
