+++ blocks/ttps_test.go
@@ -3,7 +3,7 @@ package blocks_test
 import (
 	"testing"
 
-	"github.com/facebookincubator/TTP-Runner/pkg/blocks"
+	"github.com/facebookincubator/TTP-Runner/blocks"
 	"github.com/facebookincubator/TTP-Runner/pkg/logging"
 	"github.com/stretchr/testify/assert"
